+++ firmware/deimos_daq_rev7/src/board/configuring.rs
@@ -34,6 +34,7 @@ impl<'a> Board<'a> {
 
         handler!(
             systick_handler = || {
+                self.acquisition_clock_advance();
                 self.time_ns += self.dt_ns as i64;
 
                 // Poll send/recv to process incoming packets